requirements.txt
pyspark
kafka-python
pymysql
json
pandas
numpy
